Bugfix: FSS Basic and Extended List Read -a -n -l parameter fails to work as expected.
When -a, -n, and -l are used at the same time, the line position ends up being the wrong position within the file.
The -l (--line) should be relative to the --at index relative to the --name given.

This turned out to be a simple mistake where the wrong variable was being passed.