Bugfix: Infinite loop on invalid entry file and related printing problems.
The loop is not properly handling the failure state and infinitely attempts to execute a rule that will always fail.
Make sure to copy over the rule failure code to the entry status.
Make sure to exit when rule execution fails when not simulating.

The related error printing bad logic in the conditionals.
Separate the simulate and related verbosity from the non-simulate and related verbosity.