Update: Support "pid" and "show" entry settings, and bug fixes.
The pid file may be created before the filesystem is ready and it may not be deleted when the system shuts down.
This then prevents the system from booting due to init failures.
Provide rule setting for fine tuning the control over how to handle the pid file.
Now, when running "as init", the pid entry setting by default designates to only utilize the pid file when "ready".
Previously, the pid file was always being checked for before the entry is even processed.
This prevented even waiting on the "ready" action because the entry is bailing out before even being processed.

There already is a "verbose" that prints the execution of programs.
This does not, however, start by default.
Provide a "show" setting for toggling this behavior to print some of the verbose messages when in "init" "show" mode.

Update the constant strings and related variables.

The ready behavior needs to be properly set.
As far as I can tell, the logic seemed odd, if not wrong.
Remove one of the ready checks and cleanup the ready check for the explicit ready action.

Update the prebuilt rules for terminals.
For now, there will be separate terminal files to execute.
There should be a variable substitution later on to not require multiple files.
For example, behavior is now:
  "rule terminal one", found in rules/terminal/one.rule
  "rule terminal two", found in rules/terminal/two.rule
This could better be something like:
  "rule terminal tty/1", found in rules/terminal/tty.rule
  "rule terminal tty/2", found in rules/terminal/tty.rule

Cleanup and update the documentation.