Update: add additional fake settings, simplify defines_all and flags_all.
Add the following fake build settings:
- libraries_shared
- libraries_static
- build_sources_headers_shared
- build_sources_headers_static
- build_sources_library_shared
- build_sources_library_static
- build_sources_program_shared
- build_sources_program_static
- flags_library_shared
- flags_library_static
- flags_program_shared
- flags_program_static
- defines_library_shared
- defines_library_static
- defines_program_shared
- defines_program_static

"flags_all" is now "flags".
"defines_all" is now "defines".

These are needed for the adding settings specific to shared or static state.
The "headers" related ones are added in case custom headers include static or shared specific code that should only be installed when using static or shared.
Note that the "headers" is for installation and is not needed in any way for compilation.
Other non-compilation related settings are not to have the "_shared" or "_static" added.