Cleanup: Break apart controller sources.
The functions have gotten too large an are in a need of cleanup.
Create additional source files and move functions around as appropriate.

Do some function renaming as well.

Move the print functions into separate print source files.
This is done only for functions whose sole purpose is printing and is not nor should it be done for functions that happen to have printing within them.

The entry pre-process and process functions should be in the entry source files.
Fix the name of those functions to start with "controller_entry_" rather than starting with "controller_process_".

Move any structure construct and destruct functions into the common.h and common.c that are not already there.

Break out the build settings file settings into multiple lines.
I prefer not to have too many extra lines, but there is a point when the lines get absurdly too long.
The lines are broken up by some context or pattern.