Update: Add "success" color context and initialize color "set" in programs.
There is an "error" and a "warning" but no "success".
Make this complete by adding "success".

The programs aren't initializing the "set".
Attempting to use these always results in no color context.
Fix this by initializing these in each program.

The output context also needs to be initialized just like the error and warning.